+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 [...]
+
+# v47.1.0 (05/02/2021)
+
 - **[NEW]** Add `Marquee`
 - **[NEW]** Add `TripCardSample`